requirements.txt
django
djangorestframework
environs
azure-storage-blob
azure-keyvault-secrets
openai==0.28
langchain-community
azure-cosmos
PyJWT
phonenumbers
requests
razorpay